@@ -1,6 +1,6 @@
 "use strict";
 
-module.exports = `                             _               _
+export default `                             _               _
  ___ ___ _ _ ___ ___ ___ ___| |_ ___ ___ ___| |_
 |  _| -_| | | -_|  _| -_|   |  _| . | -_| -_| '_|
 |_| |___|\\_/|___|_| |___|_|_|_| |_  |___|___|_,_|
